Added new class for rest api
